requirements.txt
django
django-filter
django-mptt
pillow